Fix off-by-one-ish bug
Increase maximum to account for the weekend.

The bug is that if a person ran the download script for the first time
on a Monday before 5pm, as I did, they wouldn't get any files.  The
datadir would have no existing files (0) and adding 2 means that the
maximum was equal to x (yesterday plus another day is two days ago), so
no download was performed.  Adding three to the maximum addresses this,
and a person running the download on a Monday with an empty datadir will
now get Friday's file.